add core.internal.Logger
