+++ android/src/main/java/com/yubico/yubikit/android/transport/usb/UsbDeviceManager.java
@@ -24,7 +24,7 @@ import android.content.IntentFilter;
 import android.hardware.usb.UsbDevice;
 import android.hardware.usb.UsbManager;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 
 import org.slf4j.LoggerFactory;
