+++ android/src/main/java/com/yubico/yubikit/android/ui/YubiKeyPromptConnectionAction.java
@@ -21,7 +21,7 @@ import android.os.Bundle;
 
 import androidx.annotation.WorkerThread;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.YubiKeyConnection;
 import com.yubico.yubikit.core.YubiKeyDevice;
 import com.yubico.yubikit.core.application.CommandState;
