+++ core/src/main/java/com/yubico/yubikit/core/fido/FidoProtocol.java
@@ -15,7 +15,7 @@
  */
 package com.yubico.yubikit.core.fido;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.Version;
 import com.yubico.yubikit.core.application.CommandState;
 import com.yubico.yubikit.core.util.StringUtils;
