+++ core/src/main/java/com/yubico/yubikit/core/otp/OtpProtocol.java
@@ -15,7 +15,7 @@
  */
 package com.yubico.yubikit.core.otp;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.Version;
 import com.yubico.yubikit.core.application.CommandException;
 import com.yubico.yubikit.core.application.CommandState;
