+++ management/src/main/java/com/yubico/yubikit/management/ManagementSession.java
@@ -16,7 +16,7 @@
 
 package com.yubico.yubikit.management;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.Transport;
 import com.yubico.yubikit.core.UsbInterface;
 import com.yubico.yubikit.core.Version;
