+++ oath/src/main/java/com/yubico/yubikit/oath/OathSession.java
@@ -16,7 +16,7 @@
 
 package com.yubico.yubikit.oath;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.Version;
 import com.yubico.yubikit.core.smartcard.AppId;
 import com.yubico.yubikit.core.application.ApplicationNotAvailableException;
