+++ piv/src/main/java/com/yubico/yubikit/piv/jca/PivCipherSpi.java
@@ -16,7 +16,7 @@
 
 package com.yubico.yubikit.piv.jca;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.util.Callback;
 import com.yubico.yubikit.core.util.Result;
 import com.yubico.yubikit.piv.KeyType;
