+++ yubiotp/src/main/java/com/yubico/yubikit/yubiotp/YubiOtpSession.java
@@ -16,7 +16,7 @@
 
 package com.yubico.yubikit.yubiotp;
 
-import com.yubico.yubikit.core.Logger;
+import com.yubico.yubikit.core.internal.Logger;
 import com.yubico.yubikit.core.Transport;
 import com.yubico.yubikit.core.Version;
 import com.yubico.yubikit.core.YubiKeyConnection;
